Prevent HK2 AbstractBinder from being configured twice. Revert the change that introduced BinderConfigurationFactory SPI. Added possibility to register HK2 AbstractBinder class along with the instance option.
