+++ inject/hk2/src/main/java/org/glassfish/jersey/inject/hk2/DelayedHk2InjectionManager.java
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2017, 2018 Oracle and/or its affiliates. All rights reserved.
+ * Copyright (c) 2017, 2019 Oracle and/or its affiliates. All rights reserved.
  *
  * This program and the accompanying materials are made available under the
  * terms of the Eclipse Public License v. 2.0, which is available at
@@ -91,6 +91,8 @@ public class DelayedHk2InjectionManager extends AbstractHk2InjectionManager {
     public void register(Object provider) throws IllegalArgumentException {
         if (isRegistrable(provider.getClass())) {
             providers.add((org.glassfish.hk2.utilities.Binder) provider);
+        } else if (Class.class.isInstance(provider) && isRegistrable((Class) provider)) {
+            providers.add((org.glassfish.hk2.utilities.Binder) createAndInitialize((Class) provider));
         } else {
             throw new IllegalArgumentException(LocalizationMessages.HK_2_PROVIDER_NOT_REGISTRABLE(provider.getClass()));
         }
