--- inject/hk2/src/main/resources/META-INF/services/org.glassfish.jersey.inject.spi.BinderConfigurationFactory
@@ -1,1 +1,0 @@
-org.glassfish.jersey.inject.hk2.HK2BinderConfigurationFactory
